Reorganize module structure
