--- src/server/mod.rs
@@ -1,12 +1,0 @@
-/// Server-side state management.
-pub mod state;
-
-/// gRPC service implementation.
-pub mod service;
-
-/// Server configuration and rate limiting.
-pub mod config;
-
-pub use config::{RateLimiter, ServerConfig};
-pub use service::AuthServiceImpl;
-pub use state::ServerState;
