+++ src/verifier/service.rs
