+++ src/verifier/state.rs
